llm_recommender/requirements.txt
# Web framework
fastapi==0.110.0
sqlalchemy==0.29.0

# ORM and database
sqlalchemy==2.0.28
asyncpg==0.29.0            # For PostgreSQL
alembic==1.13.1            # DB migrations

# Pydantic
pydantic==2.6.4
pydantic-settings==2.2.1

# Authentication and security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Caching
aiocache==0.12.1
redis==5.0.3

# LLM and NLP
transformers==4.40.0       # Hugging Face models
torch==2.2.2               # Needed for transformers
sentence-transformers==2.6.1  # Embedding + similarity

# Monitoring & logging
loguru==0.7.2
httpx==0.27.0              # Async HTTP client

# Utilities
python-dotenv==1.0.1
tqdm==4.66.2
numpy==1.26.4
scikit-learn==1.4.2

# Testing
pytest==8.1.1
pytest-asyncio==0.23.6     # (httpx already included above)